\name{sparcl-internal}
\title{Internal sparcl functions}
\alias{print.hclustsparse}
\alias{plot.hclustsparse}
\alias{print.kmeanssparse}
\alias{plot.kmeanssparse}
\alias{print.kmeanssparseperm}
\alias{plot.kmeanssparseperm}
\alias{print.hclustsparseperm}
\alias{plot.hclustsparseperm}
\alias{print.multikmeanssparse}
\alias{plot.multikmeanssparse}
\description{Internal sparcl functions}
\usage{
plot.hclustsparse(x,...)
print.hclustsparse(x,...)
plot.kmeanssparse(x,...)
print.kmeanssparse(x,...)
plot.kmeanssparseperm(x,...)
print.kmeanssparseperm(x,...)
plot.multikmeanssparse(x,...)
print.multikmeanssparse(x,...)
plot.hclustsparseperm(x,...)
print.hclustsparseperm(x,...)
}
\author{Daniela M. Witten and Robert Tibshirani}
\keyword{internal}




