\name{sparcl-package}
\alias{sparcl-package}
\alias{sparcl}
\docType{package}
\title{
Performs sparse hierarchical and sparse K-means clustering
}
\description{
Implements the sparse clustering methods of Witten and Tibshirani (2010)
"A framework for feature selection in clustering", Journal
Amer. Stat. Assocn. 105(490): 713-726.
}
\details{
\tabular{ll}{
Package: \tab sparcl\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2011-12-06\cr
License: \tab GPL >=2 \cr
LazyLoad: \tab yes\cr
}
The main functions are KMeansSparseCluster and
HierarchicalSparseCluster. Tuning parameters for these methods are
chosen by KMeansSparseCluster.permute and HierarchicalSparseCluster.permute.
}
\author{
  Daniela M. Witten and Robert Tibshirani
  
Maintainer: dwitten at stanford dot edu 
}
\references{
Witten and Tibshirani (2010) A framework for feature selection in
clustering. Journal
Amer. Stat. Assocn. 105(490): 713-726.
}
